requirements.txt
# --- LangChain core y ecosistema ---
langchain>=0.2,<0.3
langchain-community>=0.2,<0.3
langchain-openai>=0.2,<0.3
langchain-huggingface>=0.1,<0.2
langgraph>=0.2,<0.3

# Vector store
faiss-cpu>=1.7.4

# OpenAI SDK (usado por langchain-openai / AzureChatOpenAI)
openai>=1.40,<2

# --- FastAPI y utilidades web ---
fastapi>=0.111,<0.116
uvicorn[standard]>=0.23,<0.31
python-dotenv>=1.0.1,<2

# --- NLP / ML / Embeddings ---
sentence-transformers>=2.7,<3
torch>=2.2,<3
scikit-learn>=1.3,<1.6
numpy>=1.26,<3

# --- Gráficos / grafos ---
plotly>=5.20,<6
networkx>=3.2,<4

# --- PDF / utilidades varias ---
pymupdf>=1.24,<1.26
tqdm>=4.66,<5
